[ADD]opd-5682 采购申请跨年预算新建逻辑 2021-11-03 王杰

+++ src/main/java/org/srm/purchasecooperation/order/domain/entity/RcwlBudgetDistribution.java
@@ -68,7 +68,7 @@ public class RcwlBudgetDistribution extends AuditDomain {
     private Long prLineId;
     @ApiModelProperty(value = "预算占用年份",required = true)
     @NotNull
-    private Long budgetDisYear;
+    private Integer budgetDisYear;
     @ApiModelProperty(value = "预算占用金额（四舍五入）",required = true)
     @NotNull
     private BigDecimal budgetDisAmount;
